WIP batch transaction validation refactor
